refactor: moved non-api code to core
