+++ api_swedeb/api/utils/corpus.py
@@ -10,10 +10,10 @@ from penelope.corpus import VectorizedCorpus
 from penelope.corpus.dtm.interface import IVectorizedCorpus
 from penelope.utility import PropertyValueMaskingOpts  # type: ignore
 
-from api_swedeb.api.parlaclarin import codecs as md
-from api_swedeb.api.parlaclarin import speech_text as sr
-from api_swedeb.core.trends_data import SweDebComputeOpts, SweDebTrendsData
 from api_swedeb.api.utils.protocol_id_format import format_protocol_id
+from api_swedeb.core import codecs as md
+from api_swedeb.core import speech_text as sr
+from api_swedeb.core.trends_data import SweDebComputeOpts, SweDebTrendsData
 from api_swedeb.core.utility import Lazy
