+++ api_swedeb/core/__init__.py
@@ -1,5 +1,5 @@
 # type: ignore
 
-from .cwb import to_cqp_exprs, to_cqp_pattern
 from .configuration import Config, ConfigStore, ConfigValue
-from trends_data import SweDebComputeOpts, SweDebTrendsData
+from .cwb import to_cqp_exprs, to_cqp_pattern
+from .trends_data import SweDebComputeOpts, SweDebTrendsData
