+++ api_swedeb/core/n_grams/compute.py
@@ -5,7 +5,8 @@ from ccc import Corpus, SubCorpus
 
 from api_swedeb.core.cwb import to_cqp_exprs
 
- # pylint: disable=redefined-outer-name
+# pylint: disable=redefined-outer-name
+
 
 def to_n_grams(phrase: str, n: int) -> Iterable[str]:
     """Generate n-grams from a given phrase."""
